backend/requirements.txt
# API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Base de données
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache & sessions
redis==5.0.4
hiredis==2.3.2

# Auth & sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# SumUp (pas de SDK officiel — on utilise l'API REST via httpx)
# httpx déjà inclus ci-dessous

# Google Gemini API
google-generativeai==0.7.2

# Brevo (emails)
sib-api-v3-sdk==7.6.0

# Google Calendar
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0

# CalDAV
caldav==1.3.9
vobject==0.9.6.1

# Utilitaires
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1

# Génération PDF (factures)
weasyprint==62.1

# Tâches planifiées
apscheduler==3.10.4
bcrypt==4.1.3
